relay_control/requirements.txt
Flask
RPi.GPIO
smbus2